readme
##Due to the recent political events in the country DDUU, the Government has decided to contribute to the preservation of democracy, for which the government made an app that consults citizens regarding various questions and decisions. In this order, decisions are made according to popular consensus.

All users can sign up, just logged users can answer the questions. Every question will have different options. Questions will have at least one of the options as an answer. When a question is answered, the user will have a view with the statistics for that question. Once a question is answered, the question can not be changed.